Digirusを最新バージョンに更新しました

+++ Digirus (2024.6-2024.9)/プロジェクトファイル/Digirus/src/general/benlib.cpp
@@ -52,7 +52,7 @@ float DegreeToRadian(const int& nDegree)
 
 int RadianToDegree(const float& fRadian)
 {
-	return fRadian * 180.0f / (float)D3DX_PI;
+	return static_cast<int>(fRadian * 180.0f / (float)D3DX_PI);
 }
 
 D3DXVECTOR3 QuaternionToEuler(D3DXQUATERNION q)
